+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 5.0.3 (In development)
 
+* **[Feature]** Add setDataResidencyRegion(string dataResidencyRegion)" method to portal as property for Analytics and Crashes.
+
 ## Version 5.0.2
 
 ### AppCenter