+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### App Center Distribute
 
-* **[Fix]** Add RECEIVER_EXPORTED flag for install receiver
-* **[Fix]** Add FLAG_ALLOW_UNSAFE_IMPLICIT_INTENT flag for broadcast pending intent
+* **[Fix]** Add RECEIVER_EXPORTED flag for install receiver.
+* **[Fix]** Add FLAG_ALLOW_UNSAFE_IMPLICIT_INTENT flag for broadcast pending intent.
 
 ## Version 5.0.2