+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 * **[Fix]** Add RECEIVER_EXPORTED flag for install receiver.
 * **[Fix]** Add FLAG_ALLOW_UNSAFE_IMPLICIT_INTENT flag for broadcast pending intent.
 
+## Version 5.0.3
+
+### AppCenter
+
+* **[Internal]** Add `dataResidencyRegion` option.
+
+
 ## Version 5.0.2
 
 ### AppCenter